requirements.txt
openai==0.28.1
azure-storage-blob
pandas
openpyxl
streamlit
python-docx
python-dotenv